feat(wallet-sdk): own the Supabase client + RLS session-token provider (connection-core step 3)
The SDK now owns the RLS-scoped Supabase connection — the third connection-core piece on the
path to backing the accounts domain with SDK-owned connections.

- src/supabase-session.ts (moved verbatim from agicash-db/supabase-session.ts): the
  opensecret generateThirdPartyToken query with its dynamic JWT-exp staleTime, plus the
  module-load getQueryClient() capture — which now resolves INSIDE the SDK (same singleton,
  no more cross-package capture). FetchQueryOptions now typed via @tanstack/query-core.
- src/auth.ts (moved verbatim from features/shared/auth.ts): isLoggedIn — it reads the token
  keys the OpenSecret client writes, so it is opensecret-session logic; its direct
  window.localStorage read is folded into the same tracked exception as the opensecret
  dependency (switches to the StorageAdapter on the storage-pluggable bump).
- src/agicash-db.ts (new factory): createAgicashDb({url, anonKey}) — the createClient call
  moved verbatim (schema 'wallet', session-token accessToken, the production payload-redacting
  realtime logger). The consumer supplies connection params.
- web's database.client.ts keeps what is genuinely web: import.meta.env reading, the dev-LAN
  URL rewrite (window.location), the SupabaseRealtimeManager and the window debug attach —
  and now builds agicashDbClient via the SDK factory. shared/auth.ts +
  agicash-db/supabase-session.ts are transitional re-export shims (all consumers untouched;
  _protected.tsx's direct supabase-session import covered).
- SDK gains @supabase/supabase-js (catalog) + jwt-decode (hoisted to catalog, now shared)

typecheck, biome, build all green; 138 tests; no direct react/DOM imports in SDK src beyond
the documented opensecret/localStorage exception.
