+++ apps/web-wallet/app/features/agicash-db/database.client.ts
@@ -1,7 +1,5 @@
-import { createClient } from '@supabase/supabase-js';
+import { createAgicashDb } from '@agicash/wallet-sdk/agicash-db';
 import { SupabaseRealtimeManager } from '~/lib/supabase';
-import type { Database } from './database';
-import { getSupabaseSessionToken } from './supabase-session';
 
 const getSupabaseUrl = () => {
   const supabaseUrl = import.meta.env.VITE_SUPABASE_URL ?? '';
@@ -35,41 +33,10 @@ if (!supabaseAnonKey) {
  * The client-side Supabase database client.
  * If you need to use a client on the server, which bypasses RLS, use `agicashDbServer` instead.
  */
-export const agicashDbClient = createClient<Database>(
-  supabaseUrl,
-  supabaseAnonKey,
-  {
-    accessToken: getSupabaseSessionToken,
-    db: {
-      schema: 'wallet',
-    },
-    realtime: {
-      logger: (kind: string, msg: string, data?: unknown) => {
-        const now = Date.now();
-        const logData: Record<string, unknown> = {
-          timestamp: now,
-          time: new Date(now).toISOString(),
-          data,
-        };
-        if (
-          process.env.NODE_ENV === 'production' &&
-          kind === 'receive' &&
-          typeof logData.data === 'object' &&
-          logData.data != null &&
-          'payload' in logData.data
-        ) {
-          // We don't want this to log the app data for receive messages in production.
-          logData.data = {
-            ...logData.data,
-            payload: '<redacted>',
-          };
-        }
-        console.debug(`Realtime ${kind}: ${msg}`, logData);
-      },
-      logLevel: 'info',
-    },
-  },
-);
+export const agicashDbClient = createAgicashDb({
+  url: supabaseUrl,
+  anonKey: supabaseAnonKey,
+});
 
 /**
  * The client-side Supabase realtime client.
