+++ apps/web-wallet/package.json
@@ -60,7 +60,7 @@
     "embla-carousel-react": "8.5.2",
     "express": "5.2.1",
     "isbot": "5.1.34",
-    "jwt-decode": "4.0.0",
+    "jwt-decode": "catalog:",
     "ky": "1.14.3",
     "light-bolt11-decoder": "3.2.0",
     "lucide-react": "0.468.0",
