+++ package.json
@@ -18,6 +18,7 @@
       "@types/bun": "1.3.11",
       "big.js": "7.0.1",
       "dotenv": "16.4.7",
+      "jwt-decode": "4.0.0",
       "jwt-encode": "1.0.1",
       "type-fest": "5.4.3",
       "typescript": "5.9.3",
